fix(temporal): define bulk spend-limit workflow on the metronome worker
The bulk spend-limit workflow/activity were re-exported into the
metronome_events_queue worker. The workflow re-export used the `@app/*` alias,
which the Temporal workflow bundler (webpack) can't resolve for runtime imports,
so the worker bundle failed to compile and the workflow never ran.

Move the workflow and activity definitions into the metronome worker's own
workflows/activities modules (where they're bundled and registered) instead of
re-exporting them — removing the indirection and the alias gotcha entirely. The
thin launcher (client/helpers) stays under bulk_spend_limit so the endpoint
import is unchanged; the activity test moves alongside the activity.
