+++ front/temporal/bulk_spend_limit/client.ts
@@ -2,8 +2,8 @@ import { generateRandomModelSId } from "@app/lib/resources/string_ids_server";
 import { getTemporalClientForFrontNamespace } from "@app/lib/temporal";
 import logger from "@app/logger/logger";
 import { makeBulkSpendLimitWorkflowId } from "@app/temporal/bulk_spend_limit/helpers";
-import { bulkSetUserSpendLimitWorkflow } from "@app/temporal/bulk_spend_limit/workflows";
 import { QUEUE_NAME } from "@app/temporal/metronome_events_queue/config";
+import { bulkSetUserSpendLimitWorkflow } from "@app/temporal/metronome_events_queue/workflows";
 import type { UserSpendLimit } from "@app/types/api/users/spend_limit";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
